Client Cannot Join Before Host
There was an issue with the game that if the client joined the game before the host, it all broke? I dont know why this happened since by the time theyre at champ select, theyve both joined as host client so the client selecting their champ first shouldn't have been an issue, but it was, but it is fixed.