Bullets Spawning Properly
Bullets are Spawning Properly but are not being destroyed/registered by the client. I think there is an issue with the collision detector and the check if the collision.gameObject is the same as the target player. May change the target to be the parent, rather than the controller. For now, I will leave the bullets be since I want to focus on the abilities working.