Dash Works (But Not on Client)
Dash works for host, but not for client. Thinking its the same issue as for the rotation where ill just have to call it twice. Will do some digging and see if I can do a rpc specified in params thing where I can call the Rpc twice to send to self and then to everyone else.