In Game UI Mostly Finished [I'd Say It's Done] (#98)
* Health/Mana Bar Created

Working on the code to actually update it. Should be similar to the augment ones.

* UI Shows on Client, but Not Host.

* Almost there

Instead of finding the player, it finds the game manager....

* IT WORKS ON BOTH BUT

But the host takes the same dmg as the client.

* Method 1

Calling from client -> server -> client -> UI

* Update GameManager.cs

* Fixing!

* Health and Mana Bars Work!

* Clean Up

* Stretch to Screen Size

* Ability Icons Per Champion

* Working on Having the Ability Icons Change

Did have it almost working, but it had them backwards.

* TO FIX:

Icons not being set on client properly. Ashe empowered icon not working (mightve fixed it havent tested).

* Fix fix

* Hovering Over Button Works

Time to do the UI PopUp for it to explain abilities.

* Added Cooldown Logic to Ashe Q

* Timer Font and Pop Ups

The pop ups might appear.

* Hover Over Ability to See Info (#95)

* Hovering Over Button Works

Time to do the UI PopUp for it to explain abilities.

* Added Cooldown Logic to Ashe Q

* Timer Font and Pop Ups

The pop ups might appear.

* test test

* Working on Setting Icons to Buttons

* Icons Work

* Issue: Ability Icon Change and Info Box DO NOT WORK on client

Maybe I need an RPC?

* Different Results

Different results if the game is being played inside or outside of the development software

* Each Client Sets the Icons

PERFECT. Now, I need to do kind of the same thing with the mana/cooldown checks.

* Fix Interactable Logic

* Icons/Text Works

Need to fix the interactable stuff for the abilities.

* Ability Icons Work!

* Fixed Ashe Ability Icon

Ashe's Q finally updates To Show the right sprite when available.